Update according the new names, the targets: Emulator, (arduino backend devices)

+++ targets/Arduino/Arduino_Nano/examples/Nano_ADC_LEDS/Nano_ADC_LEDS.v
@@ -10,6 +10,6 @@ port.setup(port.d, 0b00000000) // all outputs
 
 for {
 	analog := adc.read(adc.ch0)
-	port.write(port.d, analog >> 2)
+	port.write(port.d, u8(analog >> 2))
 	time.sleep_ms(100)
 }
